refactor(example): move ERC-20 helpers into erc20 module
Extract find_balance_slot, find_allowance_slot, and their private helpers
(balance_slot_at, allowance_slot_at, state_override_single) into
examples/tutorial/erc20.rs. The sol! IERC20 interface and MAX_PROBE_SLOT
constant move with them. main.rs is now focused on the swap flow only.
